requirements.txt
fastapi
uvicorn
openai-whisper
torch
transformers
datasets
scikit-learn
pandas
numpy
matplotlib
seaborn
gtts
soundfile
python-multipart
loguru
jiwer